+++ CHANGELOG.md
@@ -23,6 +23,8 @@ and this project adheres to
 
 - Fix duplicate credential type bug
   [#2100](https://github.com/OpenFn/lightning/issues/2100)
+- Ensure Global OAuth Clients Accessibility for All Users
+  [#2114](https://github.com/OpenFn/lightning/issues/2114)
 
 ## [v2.5.0-pre3] - 2024-05-20